requirements.txt
python-telegram-bot[job-queue]==21.6

psycopg2-binary==2.9.9

requests==2.32.5
beautifulsoup4==4.14.3

python-dotenv==1.2.1

APScheduler==3.10.4
tzlocal==5.3.1
pytz==2025.2

httpx==0.27.2
anyio==4.12.1
h11==0.16.0
sniffio==1.3.1
typing_extensions==4.15.0

certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
six==1.17.0
soupsieve==2.8.1